tsne-backend/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
checksumdir==1.2.0
click==8.0.3
cycler==0.11.0
flake8==4.0.1
Flask==2.0.3
Flask-RESTful==0.3.9
fonttools==4.29.1
importlib-metadata==4.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
